[alpakatest] Support serial, TBB, CUDA and ROCm at the same time
Split compilation by backend:
  - the ALPAKA_..._ENABLED macros are only define done at a time;
  - introduce new ALPAKA_..._PRESENT macros to identify all backends for
    which support is being compiled;
  - link all libraries to the main executable, including the "portable",
    backend-specific ones.

Add explicit instantiation definitions and declarations to the
initialisation code, and move it to the AlpakaCore "portable" library.
